+++ tx-backend/src/integration/groovy/org/eclipse/tractusx/traceability/common/support/AssetTestData.groovy
@@ -31,7 +31,7 @@ class AssetTestData {
 
     List<Asset> readAndConvertAssetsForTests() {
         try {
-            InputStream file = AssetTestData.class.getResourceAsStream("/data/irs_assets_v2.json")
+            InputStream file = AssetTestData.class.getResourceAsStream("/data/irs_assets_v3.json")
             JobDetailResponse response = mapper.readValue(file, JobDetailResponse.class)
             def assets = response.convertAssets()
             return assets