+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/model/response/relationship/Aspect.java
@@ -25,8 +25,8 @@ import java.util.List;
 
 public enum Aspect {
     BATCH("Batch"),
-    SERIAL_PART_TYPIZATION("SerialPartTypization"),
-    ASSEMBLY_PART_RELATIONSHIP("AssemblyPartRelationship"),
+    SERIAL_PART("SerialPart"),
+    SINGLE_LEVEL_BOM_AS_BUILT("SingleLevelBomAsBuilt"),
     SINGLE_LEVEL_USAGE_AS_BUILT("SingleLevelUsageAsBuilt"),
     SINGLE_LEVEL_BOM_AS_PLANNED("SingleLevelBomAsPlanned"),
     PART_SITE_INFORMATION_AS_PLANNED("PartSiteInformationAsPlanned"),
@@ -44,11 +44,11 @@ public enum Aspect {
     }
 
     public static List<String> downwardAspectsForAssetsAsBuilt() {
-        return List.of(BATCH.getAspectName(), SERIAL_PART_TYPIZATION.getAspectName(), ASSEMBLY_PART_RELATIONSHIP.getAspectName());
+        return List.of(BATCH.getAspectName(), SERIAL_PART.getAspectName(), SINGLE_LEVEL_BOM_AS_BUILT.getAspectName());
     }
 
     public static List<String> upwardAspectsForAssetsAsBuilt() {
-        return List.of(BATCH.getAspectName(), SERIAL_PART_TYPIZATION.getAspectName(), SINGLE_LEVEL_USAGE_AS_BUILT.getAspectName());
+        return List.of(BATCH.getAspectName(), SERIAL_PART.getAspectName(), SINGLE_LEVEL_USAGE_AS_BUILT.getAspectName());
     }
 
     public static List<String> downwardAspectsForAssetsAsPlanned() {